ModloaderMP start and a bunch of tweak to project
Refmap is completely broken and I have no idea why.
There are probably a bunch of broken methods and stuff. Too bad. I don't know where or why.
I cannot do shockahpi because it completely replaces ItemTool which is dumb and stupid.

Consider me abandoning this project for the next month because I am completely done with how this crap is laid out.